Issue 155 pt.3: Streamlined JPAWebloggerImpl.
