+++ app/src/main/java/org/apache/roller/weblogger/ui/rendering/plugins/comments/AkismetCommentValidator.java
@@ -78,7 +78,7 @@ public class AkismetCommentValidator implements CommentValidator {
             URLConnection conn = url.openConnection();
             conn.setDoOutput(true);
 
-            conn.setRequestProperty("User_Agent", "Roller " + WebloggerFactory.getWeblogger().getVersion()); 
+            conn.setRequestProperty("User_Agent", "Roller " + WebloggerConfig.getProperty("weblogger.version", "Unknown"));
             conn.setRequestProperty("Content-type", "application/x-www-form-urlencoded;charset=utf8"); 
             conn.setRequestProperty("Content-length", Integer.toString(sb.length()));
