+++ app/src/main/resources/org/apache/roller/weblogger/config/roller.properties
@@ -483,9 +483,10 @@ log4j.logger.com.danga.MemCached=ERROR
 # Other settings
 #-----------------------------------------------------------------------------
 
-# Values of the properties in this list get system property expansion 
-# applied to them when loaded.
-config.expandedProperties=search.index.dir,mediafiles.storage.dir
+# populated by Maven during build
+weblogger.version=${project.version}
+weblogger.revision=r${buildNumber}
+weblogger.buildTime=${timestamp}
 
 # True to enable history in Atom feeds with next/prev links
 site.newsfeeds.history.enabled=false
