--- app/src/main/resources/roller-version.properties
@@ -1,5 +1,0 @@
-ro.version=${project.version}
-ro.revision=r${buildNumber}
-ro.buildTime=${timestamp}
-ro.buildUser=${user.name}
-
